+++ ceno_zkvm/src/scheme/hal.rs
@@ -4,11 +4,14 @@ use crate::{
     structs::{ProofInput, TowerProofs},
 };
 use ff_ext::ExtensionField;
-use mpcs::Point;
+use mpcs::{Point, PolynomialCommitmentScheme};
+use rkyv::Serialize;
+use serde::de::DeserializeOwned;
 use transcript::Transcript;
 
 pub trait ProverBackend {
     type E: ExtensionField;
+    type PcsOpeningProof: Clone + Serialize + DeserializeOwned;
     type Matrix: Send + Sync + Clone;
     type MmcsProverData;
     type MultilinearPoly: Send + Sync;
@@ -29,13 +32,15 @@ pub trait TowerProver<PB: ProverBackend> {
     // and then build a complete binary tree to accumulate these records
     fn build_tower_witness(
         &self,
-        input: ProofInput<PB::E>,
+        input: ProofInput<PB>,
         read_exprs: &[Expression<PB::E>],
         write_exprs: &[Expression<PB::E>],
         lookup_exprs: &[Expression<PB::E>],
         challenge: &[PB::E; 2],
     ) -> (Vec<TowerProverSpec<PB>>, Vec<TowerProverSpec<PB>>);
 
+    // the validity of value of first layer in the tower tree is reduced to
+    // the validity of value of last layer in the tower tree through sumchecks
     fn prove_tower_relation(
         &self,
         prod_specs: Vec<TowerProverSpec<PB>>,
@@ -46,13 +51,29 @@ pub trait TowerProver<PB: ProverBackend> {
 }
 
 pub trait MainSumcheckProver<PB: ProverBackend> {
+    // this prover aims to achieve two goals:
+    // 1. the validity of last layer in the tower tree is reduced to
+    //    the validity of read/write/logup records through sumchecks;
+    // 2. multiple multiplication relations between witness multilinear polynomials
+    //    achieved via zerochecks.
     fn prove_main_constraints(
         &self,
+        rt_tower: Vec<PB::E>,
+        tower_proof: &TowerProofs<PB::E>,
+        r_records: Vec<PB::MultilinearPoly>,
+        w_records: Vec<PB::MultilinearPoly>,
+        lk_records: Vec<PB::MultilinearPoly>,
         input: ProofInput<PB>,
-        tower_proof: TowerProofs<PB::E>,
         cs: ConstraintSystem<PB::E>,
         transcript: &mut impl Transcript<PB::E>,
-    ) -> (Point<PB::E>, TowerProofs<PB::E>);
+    ) -> Point<PB::E>;
 }
 
-pub trait OpeningProver<PB: ProverBackend> {}
+pub trait OpeningProver<PB: ProverBackend> {
+    fn open(
+        &self,
+        inputs: Vec<ProofInput<PB>>,
+        points: Vec<Point<PB::E>>,
+        transcript: &mut impl Transcript<PB::E>,
+    ) -> PB::PcsOpeningProof;
+}
